Move the SessionNavHostFragment to the navigation.session package
