+++ core/src/main/kotlin/dev/hotwire/core/turbo/delegates/TurboNestedFragmentDelegate.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import androidx.annotation.IdRes
 import androidx.fragment.app.Fragment
 import androidx.navigation.NavController
+import dev.hotwire.core.navigation.session.SessionNavHostFragment
 import dev.hotwire.core.turbo.nav.HotwireNavDestination
-import dev.hotwire.core.turbo.session.SessionNavHostFragment
 import dev.hotwire.core.turbo.visit.VisitOptions
 
 /**
